Allow `Instructor` instances use `Schedule` objs
Access to online schedule data was encapsulated by an `Instructor()` instance. This design required `source` to be specified
once per `Instructor()` instance. It is more likely that the information needed for all `Instructor()` instances is given in a single
reference to `source`.

`Instructor().__init__(...)` was changed to take a `Schedule()` object. So in `main()`, a single `Schedule()` instance is 
instantiated and passed to each `Instructor()` instance.