Allow app to access Google Sheets
Downloading a CSV file each day is vulnerable to human error.
However, without access to the Google Sheet, manual downloads
are the only way to parse schedule data.

By adding this module, the app itself is authorized to
access Google Drive/Sheets via the name of the sheet.
As a result, parsing schedule data can be scheduled
progromatically and the notification process can be fully
automated.