Generalize local dataframe area search
Under the assumption that all instructors have unique names, at most
an instructor is responsible for all clients within 3 rows of their
rowID (above, at, and below). So, calculating the rowIDs means +/- 1.

This calculation fails if the instructor's name repeats, since 2
`Instructor()` objects with the same name are the same object.
The calculation for `area` is now based on the range spanned by the
minimum and maximum rowIDs where the repeated name appears.