Artificially append a row to fetched GoogleSheet
Before, preprocessing the GoogleSheet did not account for repeated
instructors' names. So, calculating the effective area for each
instructor needed to change.

One edge case of repeated names happens when the repeated names
occur at the end of the online schedule. In this case, the parser
for GoogleSheet objects is conservative; fetches only the minimum
number of rows to include all the data. So, calculating
the effective area would raise a KeyError in the edge case.

Adding a blank row to the dataframe in preprocessing will eliminate
the KeyError.